minor - moving components before final reorg
